Create brightness.py
A brightness control script for OLED displays which do not have a backlight to control with xbacklight. The script will search for the primary screen and control the brightness with xrandr. The value of the brightness is stored in a file.

The script can be used for several things:
- run it with +value to up the brightness
- run it with -value to lower the brightness
- run it with value to set the brightness
The resulting brightness value must be between 0.0 and 1.0

The script returns the current value of the brightness (between 0.0 and 1.0) so you can use the script also without parameter to show the value on a status bar like i3blocks.